Add files via upload
•	Requirements
Q1. Which product categories contribute the most to sales, and how do the sales volumes of these categories fluctuate over time? Furthermore, in which specific month are peak sales observed?

R1. To understand product category sales contribution and identify peak sales periods, the User needs to visualise sales data through a pie chart for category-wise sales breakdown, a stacked bar chart for monthly category sales from January 2017 to September 2018, and an area chart for total sales volume in the same period, effectively highlighting peak sales months.

Q2. How does the type of store location influence the count of stores and the sales performance of different product categories? Additionally, what is the trend in the count of stores opening each year according to location type, and which are the top-performing stores in each location type?"

R1. To answer this question, the user needs to visualise the distribution and performance of stores based on their location types. This includes a combined bar and line graph depicting store counts and corresponding sales, a bar chart showcasing the trend of store openings by year, and a highlight table representing the sales performance of product categories per location type.

R2. The second requirement entails identifying and visualizing top-performing stores across different location types in terms of units sold.